requirements.txt
mariadb
schedule
python-dotenv